+++ game/figures/Bishop/Bishop.hpp
@@ -8,7 +8,7 @@ class Bishop : public Figure
         using Figure::Figure;
         ~Bishop();
 
-        virtual std::vector<sf::Vector2f> get_possible_moves();
+        virtual std::vector<sf::Vector2i> get_possible_moves();
         virtual bool is_current_move(sf::Vector2f mouse_clicked_pos);
 
 };
