+++ game/figures/King/King.cpp
@@ -2,5 +2,5 @@
 
 King::~King(){}
 
-std::vector<sf::Vector2f> King::get_possible_moves(){}
+std::vector<sf::Vector2i> King::get_possible_moves(){}
 bool King::is_current_move(sf::Vector2f mouse_clicked_pos){}
