+++ game/figures/Queen/Queen.hpp
@@ -10,7 +10,7 @@ class Queen : public Figure
         ~Queen();
 
 
-        virtual std::vector<sf::Vector2f> get_possible_moves();
+        virtual std::vector<sf::Vector2i> get_possible_moves();
         virtual bool is_current_move(sf::Vector2f mouse_clicked_pos);
 
 };
