+++ game/figures/Rook/Rook.cpp
@@ -3,5 +3,8 @@
 Rook::~Rook(){}
 
 
-std::vector<sf::Vector2f> Rook::get_possible_moves() {}
+std::vector<sf::Vector2i> Rook::get_possible_moves()
+{
+    
+}
 bool Rook::is_current_move(sf::Vector2f mouse_clicked_pos) {}
