+++ game/figures/Rook/Rook.hpp
@@ -9,7 +9,7 @@ class Rook : public Figure
         using Figure::Figure;
         ~Rook();
 
-        virtual std::vector<sf::Vector2f> get_possible_moves();
+        virtual std::vector<sf::Vector2i> get_possible_moves();
         virtual bool is_current_move(sf::Vector2f mouse_clicked_pos);
